fix: git

+++ src/Update/Git/Reset.php
@@ -8,7 +8,7 @@ use Slothsoft\Devtools\Misc\Update\UpdateInterface;
 class Reset implements UpdateInterface {
 
     public function runOn(Project $project) {
-        if ($project->chdir()) {
+        if ($project->chdir() and is_dir('.git')) {
             Utils::execute('git reset --hard');
             Utils::execute('git clean -f -d');
         }
